Fixed the syntax issue of cla.yml
This GitHub Actions workflow is designed to trigger the CLA (Contributor License Agreement) assistant action based on specific conditions during a pull request event.

Workflow Conditions:
The workflow will run if any one of the following conditions is true:

A comment is made with the text 'recheck': If a user comments 'recheck' on the pull request, this action will trigger.

A comment is made with the text 'I have read the CLA Document and I hereby sign the CLA': If a user confirms their acceptance of the CLA by commenting with this exact text, the action will trigger.

A pull request event of type 'pull_request_target' occurs: If the pull request event is a pull_request_target event (typically when a pull request is created or updated), the action will trigger.

Action Used:
cla-assistant/github-action@v2.1.3-beta: This action is responsible for verifying and processing CLA agreements from contributors.

Purpose:
This workflow ensures that contributors agree to the CLA when submitting code or making comments in a pull request. It allows for automatic checking of CLA acceptance through comments and triggers the appropriate CLA processing steps.

Signed-off-by: Ayush More <ayushmore42595@gmail.com>